requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
chromadb==0.5.3
sentence-transformers==2.7.0
google-genai>=1.0.0
slowapi==0.1.9
python-multipart==0.0.9
